# Don't execute i18n watcher on boot.